+++ aioclustermanager/exceptions.py
@@ -1,0 +1,5 @@
+class NotFoundEndpointException(Exception):
+    pass
+
+class ConflictCallException(Exception):
+    pass